% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectAD.R
\name{TrenaProjectAD}
\alias{TrenaProjectAD}
\title{Define an object of class TrenaProjectPlacneta}
\usage{
TrenaProjectAD(quiet = TRUE)
}
\value{
An object of the TrenaProjectAD class
}
\description{
Expression, variant and covariate data for the 21 IGAP loci
}
