% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectIGAP.R
\name{TrenaProjectIGAP}
\alias{TrenaProjectIGAP}
\title{Define an object of class TrenaProjectPlacneta}
\usage{
TrenaProjectIGAP(quiet = TRUE)
}
\value{
An object of the TrenaProjectIGAP class
}
\description{
Expression, variant and covariate data for the 21 IGAP loci
}
